Rename db folder
Clarify whats being handled in the db folder.
